chore: move share files to related location and fix typedoc warning
