--- lib/share/CONST.ts
@@ -1,13 +1,0 @@
-export const GCloudCDN = 'webcdn.17app.co';
-
-export const DEFAULT_AVATAR = `https://${GCloudCDN}/17live/ig-default.svg`;
-
-export const GA_DEV_ID = 'UA-85600241-2';
-export const GA_ID = 'UA-85600241-7';
-
-export const THUMBNAIL_BASE_URL =
-  'https://assets-17app.akamaized.net/THUMBNAIL_';
-
-export const RESOURCE_BASE_URL = 'https://webcdn.17app.co/campaign/vmo/';
-
-export const SUPPORT_LANGUAGES = ['zh_TW', 'en_US', 'ja', 'zh_HK'];
